+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## 2.14.0 [unreleased]
+## 2.14.0 [2026-04-23]
 
 ### BREAKING CHANGES